% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{combine_markov_models}
\alias{combine_markov_models}
\title{Combine markov models}
\usage{
combine_markov_models(model_a, model_b, weights = c(1, 1.5))
}
\arguments{
\item{model_a}{first model}

\item{model_b}{second model}

\item{weights}{numeric vector of 2 weights
defaults are \code{c(1, 1.5)}}
}
\description{
Builds a combined
markov model
}
