% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{export_markov_model}
\alias{export_markov_model}
\title{Export Markov Model}
\usage{
export_markov_model(markov_model, format = "json")
}
\arguments{
\item{markov_model}{a markov model object}

\item{format}{\itemize{
\item json - returns a json object
\item dict - returns dictionary object
}}
}
\value{
a json or dictionary object
}
\description{
Export Markov Model
}
