% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{generate_corpus}
\alias{generate_corpus}
\title{Generate a corpus from a
markov model}
\usage{
generate_corpus(markov_model, input_text)
}
\arguments{
\item{markov_model}{markov model object}

\item{input_text}{vector of input text}
}
\description{
Generate a corpus from a
markov model
}
