% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{generate_markovify_model}
\alias{generate_markovify_model}
\title{Generate Markovify model}
\usage{
generate_markovify_model(
  input_text,
  markov_state_size = 2,
  max_overlap_total = 15,
  max_overlap_ratio = 0.7,
  tries = 100
)
}
\arguments{
\item{input_text}{a \code{vector} of text}

\item{markov_state_size}{markov state to use
\code{default} is 2}

\item{max_overlap_total}{maximum overlap total
default is \code{15}}

\item{max_overlap_ratio}{maximum overlap ratio
\code{default is .7}}

\item{tries}{number over times to try to generate a sentence}
}
\value{
a markov model in Python object form
}
\description{
Generate a markov model for a
given input text and set of model
parameters
}
\examples{
\dontrun{
library(dplyr)
library(markovifyR)
data("linneman_lessons")
df <- linneman_lessons
lessons <-
df \%>\%
pull(textLesson)
markov_model <-
generate_markovify_model(
input_text = lessons,
markov_state_size = 3L,
max_overlap_total = 15,
max_overlap_ratio = .8
)
markov_model
}
}
