% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{generate_parsed_sentences}
\alias{generate_parsed_sentences}
\title{Generate parsed sentences}
\usage{
generate_parsed_sentences(markov_model)
}
\arguments{
\item{markov_model}{}
}
\value{
a \code{list}
}
\description{
This function returns the parsed input
text in sentence form
}
