% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{generate_sentence_starting_with}
\alias{generate_sentence_starting_with}
\title{Generate sentence starting with}
\usage{
generate_sentence_starting_with(markov_model, start_word)
}
\arguments{
\item{markov_model}{markov model object}

\item{start_word}{character of the word}
}
\value{
a character vector
}
\description{
This function generates a sentence
starting with the word you specify
if it is possible
}
