% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{generate_start_words}
\alias{generate_start_words}
\title{Generate start words}
\usage{
generate_start_words(markov_model, include_cumulative_distance = TRUE)
}
\arguments{
\item{markov_model}{markov model mobject}

\item{include_cumulative_distance}{if \code{TRUE} returns
cumulative distance of the words}
}
\value{
a \code{data_frame} objects
}
\description{
Generate start words
}
