% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linneman.R
\name{get_data_linneman_lessons}
\alias{get_data_linneman_lessons}
\title{Newest Linneman lessons}
\usage{
get_data_linneman_lessons()
}
\value{
a \code{data_frame}
}
\description{
Newest Linneman lessons
}
\examples{
get_data_linneman_lessons()
}
\references{
\url{asbcllc.com/reflections/peter_linneman/}
}
