% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{install_markovify}
\alias{install_markovify}
\title{Install Markovify Python module}
\usage{
install_markovify()
}
\value{
invisible
}
\description{
Install Markovify Python module
}
\examples{
 \dontrun{
 install_markovify()
}
}
