% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linneman_lessons}
\alias{linneman_lessons}
\title{Peter Linneman's life lessons}
\format{
Data frame with columns
\describe{
\item{idLesson}{Lesson Number}
\item{textLesson}{Lesson Text}
  ...
}
}
\source{
\url{https://asbcllc.com/reflections/peter_linneman/}
}
\usage{
linneman_lessons
}
\description{
List of Peter Linnemans life lessons
}
\keyword{datasets}
