% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovify.R
\name{markovify_text}
\alias{markovify_text}
\title{Markovify text}
\usage{
markovify_text(markov_model, maximum_sentence_length = NULL,
  start_words = NULL, output_column_name = "columnName", count = 50,
  tries = 100, only_distinct = TRUE, return_message = TRUE)
}
\arguments{
\item{markov_model}{a markov model produced from \code{generate_markovify_model}}

\item{maximum_sentence_length}{the maximum length of the sentence
if \code{NULL} model will produce a random sentence of any length.}

\item{start_words}{a markov model produced with a list of start word terms}

\item{output_column_name}{name of the}

\item{count}{count of the number of sentences, default \code{50}}

\item{tries}{count of the number of tries to produce a coherent sentence}

\item{only_distinct}{if \code{TRUE} returns only distinct text}

\item{return_message}{if \code{TRUE} returns a message}
}
\value{
a \code{tibble}
}
\description{
Produces text output from
a markov model based pn user specified
parameters
}
\examples{
\dontrun{
library(dplyr)
library(markovifyR)
data("linneman_lessons")
df <- linneman_lessons
lessons <- df \%>\% pull(textLesson)
markov_model <-
generate_markovify_model(
input_text = lessons,
markov_state_size = 2L,
max_overlap_total = 15,
max_overlap_ratio = .8)

markovify_text(
markov_model = markov_model,
count = 100,
maximum_sentence_length = 1500L,
output_column_name = 'linnemanBotText',
tries = 100,
only_distinct = TRUE,
return_message = TRUE
)
}
}
