% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markovifyR-package.R
\docType{package}
\name{markovifyR}
\alias{markovifyR}
\alias{markovifyR-package}
\title{Markovify wrapper for R}
\description{
Markovify wrapper for R
}
\author{
Alex Bresler (abresler@asbcllc.com)
}
